some bug fix and some features in StatisticsFragment completed

+++ app/src/main/java/com/example/runningtracker/ui/fragment/UserRegisterFragment.kt
@@ -84,8 +84,8 @@ class UserRegisterFragment : Fragment() {
             .edit()
             .putString(Constants.KEY_NAME, name)
             .putFloat(Constants.KEY_WEIGHT, weight.toFloat())
-            .putInt(Constants.KEY_HEIGHT,weight.toInt())
-            .putInt(Constants.KEY_AGE,weight.toInt())
+            .putInt(Constants.KEY_HEIGHT,weight.toFloat().toInt())
+            .putInt(Constants.KEY_AGE,weight.toFloat().toInt())
             .putBoolean(Constants.KEY_FIRST_TIME_TOGGLE, false)
             .apply()
         return  true
